feat: record dept 임시
