+++ kosta_finalpjt/src/main/resources/templates/member/memberlist.html
@@ -41,7 +41,7 @@
 			  </tr>
 			</table>
 		</form>
-         <form action="">
+         <form action="/member/test">
 			<table class="table_w100">
 				<thead class="list_line ">
 					<tr>
@@ -53,7 +53,7 @@
 				<tbody>
 					<tr class="list_line" th:if="${mlist != null }" th:each="m:${mlist }">
 						<td>
-							<input type="checkbox" class="list_ch">
+							<input name="userid" type="checkbox" class="list_ch" th:value="${m?.userid?.usernm}">
 						</td>
 						<td class="form_td" >
 							<p class="f600 list_name" th:text="${m?.userid?.usernm }"></p>
