+++ kosta_finalpjt/src/main/resources/templates/record/dept.html
@@ -3,10 +3,13 @@
 <head>
 <meta charset="UTF-8">
 <title>Insert title here</title>
+<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
 <link rel="stylesheet" th:href="@{/css/common/reset.css}">
 <link rel="stylesheet" th:href="@{/css/common/leftnav.css}">
 <link rel="stylesheet" th:href="@{/css/common/header.css}">
 <link rel="stylesheet" th:href="@{/css/record.css}">
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.4/jquery.min.js"></script>
 </head>
 
 <body>
@@ -33,19 +36,21 @@
             <div class="record_table_wrapper">
                 <table class="record_rtable">
                     <thead>
-                        <td>일자</td>
-                        <td>요일</td>
-                        <td>근무시간</td>
-                        <td>상태</td>
+                        <td>이름</td>
+                        <td>부서</td>
+                        <td>잡 레벨</td>
+                        <td>근무일</td>
+                        <td>지각 횟수</td>
+                        <td>총 근무시간</td>
                     </thead>
                     <tbody class="record_list">
                         <tr th:each="s:${list.list}">
-                            <td th:text="${s.id}">24.05.06</td>
                             <td th:text="${s.name}">목</td>
                             <td th:text="${s.deptNum}">09:00</td>
+                            <td th:text="${s.joblv}">조기퇴근</td>
                             <td th:text="${s.totalRecords}">조기퇴근</td>
                             <td th:text="${s.lateCount}">조기퇴근</td>
-                            <td th:text="${s.lateCount}">조기퇴근</td>
+                            <td th:text="${s.workTime}">조기퇴근</td>
                         </tr>
                     </tbody>
                 </table>
@@ -56,4 +61,5 @@
 
 </div>
 </body>
+<script th:src="@{/js/record_common.js}"></script>
 </html>
